GitHub Pages publish automation
GitHub Actions have been added to automatically publish to GitHub
pages. The main branch is published to the path `/staging`, release
candidates are published to `/rc-[version]`, and published releases are
published to `/[version]` and `/[latest]`.

We can customize the paths/versioning later. We will also be adding
preview builds later.