Add support for `nvm`
A `.nvmrc` file has been added to make it easier to keep track of the
recommended version of Node.js to use. We can reference this file both
in CI and in our README, so contributors and CI are more likely to be
the same.

A "Setup" section has been added to the README to explain how to setup
using `nvm`.

This helps unblock the GitHub pages publishing automation work, which
uses GitHub action configuration that relies upon this `.nvmrc` file.