Simplify publishing flow
We have automatic publishing setup now, but it's set to use different
paths for "staging" and "production" and so on. Unfortunately
Docusaurus doesn't seem to support this; it needs to know at build-time
which path it's being hosted on (the `baseUrl` configuration setting).

The publishing flow has been updated to publish directly from the root
of the `gh-pages` branch. There is no longer a separate "staging" and
"production" environment; we will have to set those up later using
additional repositories.