Fix preview publishing
Preview publishing was broken because we were trying to use a short
commit hash as the `ref` to checkout with `actions/checkout@v3`. But
`actions/checkout@v3` only supports full commit hashes.

Instead we now pass it a full hash. Unfortunately this means that the
preview build URLs will be quite a bit longer, but hopefully this is
OK. If not, we can update this to generate both hashes.